moved metrics_csv parser to aggregation.py
