+++ lir/resources/registry.yaml
@@ -66,7 +66,7 @@ metrics:
   cllr: lir.metrics.cllr
   cllr_min: lir.metrics.cllr_min
 output:
-  csv: lir.config.experiment_strategies.metrics_csv
+  csv: lir.aggregation.metrics_csv
   pav: lir.aggregation.plot_pav
   ece: lir.aggregation.plot_ece
   lr_histogram: lir.aggregation.plot_lr_histogram
